Fixes + Localization + Change design (admin)

+++ Shop.Web/shop-web/src/app/pages/admin/products/products.component.ts
@@ -4,6 +4,8 @@ import { BaseCompleteComponent } from '../../../components/base/base-complete.co
 import { takeUntil } from 'rxjs';
 import { AdminProductDataService } from '../../../services/data/admin/admin-product-data.service';
 import { IProduct } from '../../../models/interfaces/product';
+import { Converter } from '../../../common/converter';
+import { ConfirmationService } from 'primeng/api';
 
 @Component({
   selector: 'shop-products',
@@ -17,7 +19,8 @@ export class ProductsComponent extends BaseCompleteComponent implements OnInit {
   constructor(
     private _adminProductDataService: AdminProductDataService,
     private _cd: ChangeDetectorRef,
-    private _router: Router) {
+    private _router: Router,
+    private _confirmationService: ConfirmationService) {
     super();
   }
 
@@ -33,19 +36,37 @@ export class ProductsComponent extends BaseCompleteComponent implements OnInit {
     this._router.navigate(['admin/products/create']);
   }
 
-  public delete(e: Event, product: IProduct): void {
+  public tryDelete(e: Event, product: IProduct): void {
     e.stopPropagation();
-    this._adminProductDataService.delete(product.id)
-      .subscribe(() => {
-        this.load();
-      });
+    this._confirmationService.confirm({
+      message: 'Are you sure that you want to proceed?',
+      header: 'Confirmation',
+      accept: () => {
+        this._adminProductDataService.delete(product.id)
+          .pipe(takeUntil(this.__unsubscribe$))
+          .subscribe(() => {
+            this.load();
+            this._cd.detectChanges();
+          });
+      },
+      reject: () => {
+        return;
+      }
+    });
   }
 
   private load(): void {
+    this.displayService.changeStateLoadBar(true);
     this._adminProductDataService.getAll()
       .pipe(takeUntil(this.__unsubscribe$))
       .subscribe((data: IProduct[]) => {
         this.products = data;
+        for (let product of this.products) {
+          if (product.images[0]) {
+            product.images[0].smallBody = Converter.toFileSrc(product.images[0].mimeType, product.images[0].smallBody);
+          }
+        }
+        this.displayService.changeStateLoadBar(false);
         this._cd.detectChanges();
       });
   }
